requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly.express
mlxtend